feat: Espresso testing framework and a trial espresso screen to try it and implement UI tests for Home and Scanner screens

+++ app/src/androidTest/java/com/example/barcodescanner/features/barcodehistory/data/BarcodeDaoTest.kt
@@ -140,7 +140,7 @@ class BarcodeDaoTest {
     fun whenInsertMultipleCodesThenAllAreStored() = runBlocking {
         val barcodes = (1..5).map {
             BarcodeEntity(
-                id = it.toLong(),
+                id = it,
                 code = "code$it",
                 timestamp = Date(System.currentTimeMillis() + it)
             )
